Consolidate configuration and fix bugs
This change pulls in Dynaconf for better configuration management.
Instead of having a .ini for some stuff and environment variables for
others, it refactors some of the existing code to use the new method for
grabbing settings. Now, configuration is done using:

- settings.toml for non-secret configuration
- .secrets.toml for our SMTP password and discord token
- environment variables prefixed with `ANDREWBOT_`

All config file values can be overridden with environment variables.