Implement server-side rate limiting by user_id
This change implements server-side rate limiting of verification
attempts by keying the session by discord user id in addition to the
UUID. The user id is the primary key and the UUID ensures we don't leak
any information about a user.

Additionally, the db module was refactored to use REAL OOP and provide a
SessionManager class to handle all of the state.